requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
djangorestframework>=3.14.0
django-rest-knox>=4.2.0
drf-spectacular>=0.27.0
django-filter>=23.5
cryptography>=41.0.0
